requirements.txt
Flask==3.0.3
gunicorn==22.0.0
pandas==2.2.3
openpyxl==3.1.5
xlsxwriter==3.1.2
reportlab==4.2.2
matplotlib==3.9.2
seaborn==0.13.2
wordcloud==1.9.2
arabic_reshaper==0.1.4
python-bidi==0.4.2
